Put in a fix to not remove the FIFOs of named services.
CVS patchset: 755
CVS date: 2001/08/23 20:33:12

+++ heartbeat/client_lib.c
@@ -462,8 +462,10 @@ hb_api_signoff(struct ll_cluster* cinfo)
 	OurClientID = NULL;
 	(void)fclose(pi->RequestFIFO);	pi->RequestFIFO = NULL;
 	(void)fclose(pi->ReplyFIFO);	pi->ReplyFIFO = NULL;
-	(void)unlink(pi->ReplyFIFOName);
-	(void)unlink(pi->ReqFIFOName);
+	if (pi->iscasual) {
+		(void)unlink(pi->ReplyFIFOName);
+		(void)unlink(pi->ReqFIFOName);
+	}
 	pi->ReplyFIFOName[0] = EOS;
 	pi->ReqFIFOName[0] = EOS;
 	pi->SignedOn = 0;
